+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.4.4] - 2021-02-03
 ### Fixed
  - Display a proper error message when Proton installation is incomplete due to missing Steam Runtime
  - Display a proper warning when a tool manifest is empty
@@ -122,7 +122,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.4.3...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.4.4...HEAD
+[1.4.4]: https://github.com/Matoking/protontricks/compare/1.4.3...1.4.4
 [1.4.3]: https://github.com/Matoking/protontricks/compare/1.4.2...1.4.3
 [1.4.2]: https://github.com/Matoking/protontricks/compare/1.4.1...1.4.2
 [1.4.1]: https://github.com/Matoking/protontricks/compare/1.4...1.4.1